Have average support nulls for 3.2.0
